python/mongologger/requirements.txt
mysql-connector-python==8.0.24
paho-mqtt==1.5.1
pymongo